+++ tutorial/detection/face/tutorial-face-detector-live.cpp
@@ -79,7 +79,7 @@ int main(int argc, const char* argv[])
           vpDisplay::displayText(I, bbox.getTop()-10, bbox.getLeft(), "Message: \"" + face_detector.getMessage(i) + "\"", vpColor::red);
         }
       }
-      vpDisplay::displayText(I, I.getHeight()-25, 10, "Click to quit...", vpColor::red);
+      vpDisplay::displayText(I, (int)I.getHeight()-25, 10, "Click to quit...", vpColor::red);
       vpDisplay::flush(I);
       if (vpDisplay::getClick(I, false)) // a click to exit
         break;
